refactor(move): move `renderNunjucks` function from `src/index.ts` to `src/renderer.ts`
